+++ bootstrap/server.go
@@ -26,7 +26,7 @@ func Commands() *mux.Router {
 func APIVersions(w http.ResponseWriter, r *http.Request) {
 	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
 	w.Header().Set("Api-Service", "-")
-	w.Write([]byte("/api/user/v0"))
+	w.Write([]byte("/api/k3s/v0"))
 }
 
 func main() {