+++ changelog.md
@@ -17,7 +17,7 @@
 - Fix communication between pods via different mesos agents
 - Add fetch custom bootstrap script to init kubernetes
 - Change Mesos TASK ID's
-- Change the K3S Image to plain debian. The bootstrap will install K3S.
+- Change the K3S Image to plain ubuntu. The bootstrap will install K3S.
 - Add Dashbaord configuration
 - Add K3S Framework API Server into the Container